feat(table)!: remove top-level sub-components
