+++ .storybook/recipes/TabularInputs.stories.tsx
@@ -1,15 +1,6 @@
 import type { StoryObj, Meta } from '@storybook/react';
 import React from 'react';
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableRow,
-  TableHeader,
-  TableHeaderCell,
-  InputField,
-  Label,
-} from '../../src/';
+import { Table, InputField, Label } from '../../src/';
 
 export default {
   title: 'Recipes/TabularInput',
@@ -22,32 +13,32 @@ type Args = React.ComponentProps<typeof Table>;
 export const Default: StoryObj<Args> = {
   render: (args) => (
     <Table {...args}>
-      <TableHeader>
-        <TableRow>
-          <TableHeaderCell>Label</TableHeaderCell>
-          <TableHeaderCell>Field</TableHeaderCell>
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        <TableRow>
-          <TableCell className="flex items-center">
+      <Table.Header>
+        <Table.Row>
+          <Table.HeaderCell>Label</Table.HeaderCell>
+          <Table.HeaderCell>Field</Table.HeaderCell>
+        </Table.Row>
+      </Table.Header>
+      <Table.Body>
+        <Table.Row>
+          <Table.Cell className="flex items-center">
             <Label
               className="flex items-center"
               htmlFor="field-1"
               id="input-1"
               text="Input One"
             />
-          </TableCell>
-          <TableCell>
+          </Table.Cell>
+          <Table.Cell>
             <InputField
               aria-label="redundant"
               aria-labelledby="input-1"
               id="field-1"
               placeholder="click the label to highlight field"
             />
-          </TableCell>
-        </TableRow>
-      </TableBody>
+          </Table.Cell>
+        </Table.Row>
+      </Table.Body>
     </Table>
   ),
 };
