--- src/components/TableBody/index.ts
@@ -1,1 +1,0 @@
-export { TableBody as default } from './TableBody';
