--- src/components/TableCaption/index.ts
@@ -1,1 +1,0 @@
-export { TableCaption as default } from './TableCaption';
