--- src/components/TableCell/TableCell.module.css
@@ -1,15 +1,0 @@
-@import '../../design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # TABLE CELL
-\*------------------------------------*/
-
-/**
- * Table Cell
- * The `:where` pseudo class function allows easy overriding via className.
- */
-:where(.table-cell) {
-  font: var(--eds-theme-typography-body-sm);
-
-  padding: var(--eds-size-2) var(--eds-size-1);
-}
