--- src/components/TableCell/index.ts
@@ -1,1 +1,0 @@
-export { TableCell as default } from './TableCell';
