--- src/components/TableFooter/index.ts
@@ -1,1 +1,0 @@
-export { TableFooter as default } from './TableFooter';
