--- src/components/TableHeader/index.ts
@@ -1,1 +1,0 @@
-export { TableHeader as default } from './TableHeader';
