--- src/components/TableHeaderCell/TableHeaderCell.module.css
@@ -1,35 +1,0 @@
-@import '../../design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # TABLE HEADER CELL
-\*------------------------------------*/
-
-/**
- * Table header cell
- * The `:where` pseudo class function allows easy overriding via className.
- */
-:where(.table-header-cell) {
-  font: var(--eds-theme-typography-label-sm);
-  padding: var(--eds-size-1) var(--eds-size-1);
-  text-align: left;
-}
-
-/**
- * Table header cell within table body
- *
- * Adjusts the padding to match the table__cell padding for alignment.
- */
-.table-header-cell--body {
-  padding: var(--eds-size-2) var(--eds-size-1);
-}
-
-/**
- * Table header cell button
- *
- * On table header cells that are sortable, the content is wrapped by a <Button>.
- */
-.table-header-cell__sort-button {
-  gap: var(--eds-size-half);
-  text-decoration: none;
-  font-weight: var(--eds-font-weight-bold);
-}
