--- src/components/TableHeaderCell/index.ts
@@ -1,2 +1,0 @@
-export { TableHeaderCell as default, SORT_DIRECTIONS } from './TableHeaderCell';
-export type { SortDirectionsType } from './TableHeaderCell';
