--- src/components/TableRow/TableRow.module.css
@@ -1,21 +1,0 @@
-/*------------------------------------*\
-    # TABLE ROW
-\*------------------------------------*/
-
-/**
- * Table row
- */
-.table-row {
-  vertical-align: top;
-  border-bottom-width: var(--eds-theme-border-width);
-  border-bottom-style: solid;
-  border-bottom-color: var(--eds-theme-color-border-neutral-subtle);
-}
-
-/**
- * Header variant 
- * Has darker bottom border color for distinguishment.
- */
-.table-row--header {
-  border-bottom-color: var(--eds-theme-color-border-neutral-default);
-}
