--- src/components/TableRow/index.ts
@@ -1,1 +1,0 @@
-export { TableRow as default } from './TableRow';
