+++ src/index.ts
@@ -79,13 +79,6 @@ export { default as Skeleton } from './components/Skeleton';
 export { default as Slider } from './components/Slider';
 export { default as Tab } from './components/Tab';
 export { default as Table } from './components/Table';
-export { default as TableBody } from './components/TableBody';
-export { default as TableCaption } from './components/TableCaption';
-export { default as TableCell } from './components/TableCell';
-export { default as TableFooter } from './components/TableFooter';
-export { default as TableHeader } from './components/TableHeader';
-export { default as TableHeaderCell } from './components/TableHeaderCell';
-export { default as TableRow } from './components/TableRow';
 export { default as Tabs } from './components/Tabs';
 export { default as Tag } from './components/Tag';
 export { default as Text } from './components/Text';
